backend_comites/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
dj-database-url==2.1.0
Django==4.2.1
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
fonttools==4.53.1
gunicorn==21.2.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
numpy==2.0.1
packaging==23.2
pillow==10.3.0
psycopg2==2.9.6
psycopg2-binary==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
